refactored time series tests per #1892
